Enable integration tests and clean up repo structure (too nested example).
