+++ README.md
@@ -18,7 +18,7 @@ Tools and flows for developing client-side features and extensions for [SAP Web 
 - **Testing**
   * Uses [Karma Test runner](https://github.com/karma-runner/karma).
   * Uses our home-brewed APIs, which allows programmatic access to SAP Web IDE Services (Service Test Framework).
-  
+
 - **Local development server** for static resources:
   * Provides fast feedback loops.
   * Runs the bundled version of your feature locally.
@@ -28,8 +28,8 @@ Tools and flows for developing client-side features and extensions for [SAP Web 
 
 ## Requirements and Compatibility
 * A [maintained version](https://github.com/nodejs/Release) of node.js.
-  
-  
+
+
 ## Installation
 
 Run this command:
@@ -39,7 +39,7 @@ Run this command:
 ## Usage and Documentation
 
 See these examples:
-* [A feature demonstrating client-tool capabilities](https://github.com/SAP/webide-client-tools/tree/master/example/template).
+* [A feature demonstrating client-tool capabilities](https://github.com/SAP/webide-client-tools/tree/master/example).
 * [HTML version](http://sap.github.io/webide-client-tools/web/html_docs/modules/_api_d_.html).
