+++ example/.npmignore
