+++ example/karma.conf.js
