+++ example/scripts/karma_ci.js
