+++ example/src/webide-plugin-example/command/HelloWorld.js
