+++ example/src/webide-plugin-example/i18n/i18n.properties
