+++ example/src/webide-plugin-example/plugin.json
