+++ example/src/webide-plugin-example/service/Sample.js
