+++ package.json
@@ -49,12 +49,15 @@
     "typescript": "2.7.2"
   },
   "scripts": {
-    "ci_full_flow": "npm-run-all format_validate type_check lint test docs",
+    "ci_full_flow": "npm-run-all format_validate type_check lint test test_integration docs",
     "ci_full_flow_no_docs": "npm-run-all format_validate type_check lint test",
     "release": "git fetch && git rebase && npm-run-all ci_full_flow_no_docs && node scripts/release.js",
     "lint": "eslint 'src/**/*.js' 'test/**/*.js' 'example/integration_tests/**/*.js'",
     "test": "jest",
-    "test_examples": "jest --maxWorkers 1 --config example/integration_tests/jest.config.json",
+    "test_integration": "npm-run-all init_example execute_integration_tests unlink_example",
+    "init_example": "node scripts/init_example.js",
+    "execute_integration_tests": "jest --maxWorkers 1 --config test_integration/jest.config.json",
+    "unlink_example" : "cd ./example && npm unlink @sap-webide/webide-client-tools",
     "docs": "shx rm -rf docs/web/html_docs && typedoc --tsconfig tsdocsconfig.json --out docs/web/html_docs --module commonjs --name WebideClientTools --excludeExternals --includeDeclarations",
     "format": "prettier --write \"**/*.js\"",
     "format_validate": "prettier --list-different \"**/*.js\"",
